requirements.txt
Flask
Flask-Session[cachelib]
PyYAML
pyfiglet
python-dotenv
plexapi
iso-639
iso3166
requests
jsonschema